add deprecation to pixel_values and dont catch warnings
