+++ tobler/tests/test_area_interpolators.py
@@ -15,6 +15,8 @@ def datasets():
     sac1 = geopandas.read_file(sac1.get_path("sacramentot2.shp"))
     sac2 = geopandas.read_file(sac2.get_path("SacramentoMSA2.shp"))
     sac1["pct_poverty"] = sac1.POV_POP / sac1.POV_TOT
+    categories = ["cat", "dog", "donkey", "wombat", "capybara"]
+    sac1["animal"] = (categories * ((len(sac1) // len(categories)) + 1))[: len(sac1)]
 
     return sac1, sac2
 
@@ -26,10 +28,37 @@ def test_area_interpolate_singlecore():
         target_df=sac2,
         extensive_variables=["TOT_POP"],
         intensive_variables=["pct_poverty"],
-        n_jobs=1
+        categorical_variables=["animal"],
+        n_jobs=1,
     )
     assert_almost_equal(area.TOT_POP.sum(), 1796856, decimal=0)
     assert_almost_equal(area.pct_poverty.sum(), 2140, decimal=0)
+    assert_almost_equal(area.animal_cat.sum(), 32, decimal=0)
+    assert_almost_equal(area.animal_dog.sum(), 19, decimal=0)
+    assert_almost_equal(area.animal_donkey.sum(), 22, decimal=0)
+    assert_almost_equal(area.animal_wombat.sum(), 23, decimal=0)
+    assert_almost_equal(area.animal_capybara.sum(), 20, decimal=0)
+
+
+def test_area_interpolate_custom_index():
+    sac1, sac2 = datasets()
+    sac1.index = sac1.index * 2
+    sac2.index = sac2.index * 13
+    area = area_interpolate(
+        source_df=sac1,
+        target_df=sac2,
+        extensive_variables=["TOT_POP"],
+        intensive_variables=["pct_poverty"],
+        categorical_variables=["animal"],
+    )
+    assert_almost_equal(area.TOT_POP.sum(), 1796856, decimal=0)
+    assert_almost_equal(area.pct_poverty.sum(), 2140, decimal=0)
+    assert_almost_equal(area.animal_cat.sum(), 32, decimal=0)
+    assert_almost_equal(area.animal_dog.sum(), 19, decimal=0)
+    assert_almost_equal(area.animal_donkey.sum(), 22, decimal=0)
+    assert_almost_equal(area.animal_wombat.sum(), 23, decimal=0)
+    assert_almost_equal(area.animal_capybara.sum(), 20, decimal=0)
+    assert not area.isna().any().any()
 
 
 def test_area_interpolate_sindex_options():
